+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Features
+* Start a NEAR Explorer, backed by a local backend & WAMP, backed by the public testnet indexer
 
 # 0.1.0
 * Initial commit